// src/app/models/cmdInfo.ts
export class CmdInfo{

      branchName: string;
      commitMessage: string;
      errorLog: string;
      exitCode: 0;
      localRepoPath: string;
      propertiesFilePath: string;
      propertiesToUpdate: {};
      remoteRepoURL: string;
      repoUserId: string;
      repoUserPassword: string;
      sshPrivateKeyLocalPath: string;
      statusText: string;
      successLog: string
}